Orbit_core/actions/productivity_health.py
"""
AI Productivity, Health & Lifestyle Features
Steps 5-6: AI/Productivity and Health/Lifestyle
"""

from typing import Dict, List, Optional
from datetime import datetime, timedelta
import json
from pathlib import Path

class ProductivityHealthManager:
    def __init__(self, llm_client=None):
        self.llm = llm_client
        self.screen_time_data = {}
        self.health_data = {
            'water_intake': [],
            'exercise': [],
            'sleep': [],
            'breaks': []
        }
        self.task_predictions = []
        self.points = 0
        self.level = 1
    
    # ==================== TRANSLATION ====================
    
    def translate_text(self, text: str, target_language: str = 'es') -> str:
        """Translate text to target language"""
        try:
            from googletrans import Translator  # type: ignore
            
            translator = Translator()
            result = translator.translate(text, dest=target_language)
            
            return f"Translation ({target_language}): {result.text}"
        
        except ImportError:
            return "Translation requires googletrans: pip install googletrans==4.0.0-rc1"
        except Exception as e:
            return f"Translation failed: {str(e)}"
    
    def translate_speech(self, audio_file: str, target_language: str) -> str:
        """Translate speech from audio file"""
        # This would integrate STT → Translation → TTS
        return f"Speech translation to {target_language} (implementation pending)"
    
    def detect_language(self, text: str) -> str:
        """Detect language of text"""
        try:
            from googletrans import Translator  # type: ignore
            
            translator = Translator()
            detection = translator.detect(text)
            
            return f"Detected language: {detection.lang} (confidence: {detection.confidence})"
        
        except ImportError:
            return "Language detection requires googletrans"
        except Exception as e:
            return f"Detection failed: {str(e)}"
    
    # ==================== TASK PREDICTION & SMART REMINDERS ====================
    
    def predict_next_task(self, current_time: datetime = None) -> str:
        """Predict next task based on patterns"""
        if not current_time:
            current_time = datetime.now()
        
        hour = current_time.hour
        day_of_week = current_time.weekday()
        
        # Simple rule-based predictions
        predictions = {
            (0, 8): "Start your morning routine",
            (8, 12): "Work on high-priority tasks",
            (12, 13): "Take lunch break",
            (13, 17): "Continue work tasks",
            (17, 19): "Exercise or personal time",
            (19, 22): "Dinner and relaxation",
            (22, 24): "Wind down for bed"
        }
        
        for (start, end), task in predictions.items():
            if start <= hour < end:
                return f"Suggested task: {task}"
        
        return "No specific task suggested for this time"
    
    def create_smart_reminder(self, task: str, context: Dict) -> str:
        """Create reminder with AI-powered timing"""
        if self.llm:
            prompt = f"When is the best time to remind about: {task}? Context: {context}"
            suggestion = self.llm.generate(prompt)
            return f"Smart reminder created: {suggestion}"
        else:
            return f"Reminder set for: {task}"
    
    def optimize_schedule(self, tasks: List[Dict]) -> str:
        """Optimize task schedule based on priority and time"""
        if not tasks:
            return "No tasks to optimize"
        
        # Sort by priority and estimated time
        sorted_tasks = sorted(
            tasks,
            key=lambda x: (x.get('priority', 0), -x.get('duration', 0)),
            reverse=True
        )
        
        schedule = "Optimized Schedule:\n"
        current_time = datetime.now()
        
        for task in sorted_tasks:
            duration = task.get('duration', 30)
            schedule += f"- {current_time.strftime('%H:%M')}: {task['name']} ({duration}min)\n"
            current_time += timedelta(minutes=duration)
        
        return schedule
    
    # ==================== DOCUMENT GENERATION ====================
    
    def generate_report(self, data: Dict, report_type: str = 'summary') -> str:
        """Auto-generate reports from data"""
        if self.llm:
            prompt = f"Generate a {report_type} report from this data:\n{json.dumps(data, indent=2)}"
            report = self.llm.generate(prompt)
            return report
        else:
            # Simple template-based report
            report = f"=== {report_type.upper()} REPORT ===\n"
            for key, value in data.items():
                report += f"{key}: {value}\n"
            return report